refactor(cfd-2d): Decompose spalart_allmaras/mod.rs — extract model leaf module
Part of ATLAS-ARCH-007.
